backend/requirements.txt
fastapi
uvicorn[standard]
cohere
qdrant-client
asyncpg
python-dotenv
pydantic
openai-agents
flask
requests
trafilatura
